updated GUI

+++ phase2/UserInterface/DeleteBankAccountMenuController.java
@@ -45,12 +45,22 @@ public class DeleteBankAccountMenuController extends Menu implements java.io.Ser
 				selectedAccount = account;
 			}
 		}
+		if (this.masterAccessKeyIn.getText().equals(GUI.getBM().getMasterAccessKey())) {
+			this.masterAccessKeyInStatus.setText("correct master access key");
+		} else {
+			this.masterAccessKeyInStatus.setText("incorrect master access key");
+		}
 		if (!(this.userBankAccounts.getSelectionModel().isEmpty())) {
-			// angela
+			this.userBankAccountsStatus.setText(this.userBankAccounts.getValue() + " selected");
+		} else {
+			this.userBankAccountsStatus.setText("no bank account selected. try again");
+			this.endStatus.setText("");
+		}
+		if (this.masterAccessKeyInStatus.getText().equals("correct master access key") &&
+				(!(this.userBankAccounts.getSelectionModel().isEmpty()))) {
 			GUI.getBM().deleteAccount(this.user, selectedAccount.getAccountNum());
 			this.endStatus.setText("bank account deleted");
 		} else {
-			this.userBankAccountsStatus.setText("no bank account selected. try again");
 			this.endStatus.setText("");
 		}
 	}
